--- server/src/main/java/com/wap/wapor/entity/UserType.java
@@ -1,5 +1,0 @@
-package com.wap.wapor.entity;
-
-public enum UserType {
-    KAKAO, EMAIL
-}
