--- server/src/main/java/com/wap/wapor/repository/KakaoUserRepository.java
@@ -1,9 +1,0 @@
-package com.wap.wapor.repository;
-
-import com.wap.wapor.entity.User;
-import org.springframework.data.jpa.repository.JpaRepository;
-import java.util.Optional;
-
-public interface UserRepository extends JpaRepository<User, Long> {
-    Optional<User> findByIdentifier(String identifier);
-}
